feat: Add social login (Kakao/Google/Apple), remove isFullTank, add isExistModel
- Implement Kakao login with access token direct flow
- Implement Google/Apple login with JWKS JWT verification (JwksTokenVerifier)
- Unify AuthLoginRequest to single credential field
- Remove isFullTank from fueling domain (entity, DTOs, service, repository, tests, schema)
- Simplify fuel efficiency calculation: always calculate from previous record mileage
- Add isExistModel field to bike registration
- Update CLAUDE.md and Swagger descriptions

+++ src/main/java/com/bikeridediary/domain/fueling/dto/FuelingStatsResponse.java
@@ -15,6 +15,6 @@ public record FuelingStatsResponse(
         // 최근 연비 (km/L)
         BigDecimal latestFuelEfficiency,
         // 평균 리터당 가격 (원)
-        Long averagePricePerLiter
+        Double averagePricePerLiter
 ) {
 }
